+++ src/cli.js
@@ -40,6 +40,11 @@ function logCommandHelp() {
 
 
 function buildAndRun(args) {
+    /**
+     * need to manually build for ios to handle older versions:
+     * so this will need to take in a config of project, and scheme as well as device type
+     * xcodebuild -project HthuNative.xcodeproj -configuration Debug -scheme HthuNative_DEV -destination id=3787F8C6-55CD-4CF0-8A6D-AE1310D20BB0 -derivedDataPath build
+     */
     //8081 is the metro packager default, if it's running assume we already started up
     if (checkPortListening(8081)) {
         logMessage('App Already Running, Skipping Build');
@@ -109,7 +114,10 @@ function runTests() {
             }),
             stdio: 'inherit'
         };
-        let testProcess = spawn('node_modules/.bin/jest', [jestArgs], opts);
+        let testProcess = spawn('./node_modules/.bin/jest', [jestArgs], opts);
+        testProcess.on('error', (error) => {
+            console.log('Error Starting Test Process: ', error);
+        });
         testProcess.on('exit', (data) => {
             if (data == null) {
                 logMessage('Testing Aborted');