+++ test.e2e.js
@@ -1,7 +1,0 @@
-
-
-describe('Test', () => {
-    it('Runs', () => {
-        expect(true).toBeTruthy();
-    });
-});