+++ CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+- Update babel build step to only build files need need es6 transforming, e.g. setup.js
+- Update build for iOS to be use xcodebuild instead of react-native run-ios due to the need to specify project and scheme
 ## v1.0.0 - 2018-05-16
 
 ### Added